🚚 move 'get_standard_youwol_env' in environment.online_environments
