+++ youwol/environment/__init__.py
@@ -6,3 +6,4 @@ from .auto_download_thread import *
 from .config_from_module import *
 from .projects_finders import *
 from .python_dynamic_loader import *
+from .online_environments import *
